refactor(json): fold infer-schema into 'tools json schema' subcommand (#199) Restructure src/json into a multi-command program: JSON<->TOON conversion becomes the default 'convert' command, and the former standalone infer-schema tool becomes 'tools json schema' (alias 'infer-schema'). Removes src/infer-schema.
